chore(stories): split stories
