+++ examples/storybook/.storybook/config.js
@@ -1,7 +1,9 @@
 import { configure } from '@storybook/angular';
 
+const req = require.context('../src/stories/', true, /\.stories\.ts$/);
+
 function loadStories() {
-  require('../src/stories/index.ts');
+  req.keys().forEach(filename => req(filename));
 }
 
 configure(loadStories, module);
