+++ examples/storybook/src/stories/CurrentRefinements.stories.ts
@@ -1,0 +1,55 @@
+import { storiesOf } from '@storybook/angular';
+import { wrapWithHits } from '../wrap-with-hits';
+import meta from '../meta';
+
+storiesOf('CurrentRefinements', module)
+  .addDecorator(meta)
+  .add('default', () => ({
+    component: wrapWithHits({
+      template: '<ais-current-refinements></ais-current-refinements>',
+      searchParameters: {
+        disjunctiveFacetsRefinements: { brand: ['Apple', 'Samsung'] },
+        disjunctiveFacets: ['brand'],
+        numericRefinements: { price: { '>=': [100] } },
+      },
+    }),
+  }))
+  .add('with panel header', () => ({
+    component: wrapWithHits({
+      template: `
+        <ais-panel header='Current refinements'>
+          <ais-current-refinements></ais-current-refinements>
+        </ais-panel>
+      `,
+      searchParameters: {
+        disjunctiveFacetsRefinements: { brand: ['Apple', 'Samsung'] },
+        disjunctiveFacets: ['brand'],
+        numericRefinements: { price: { '>=': [100] } },
+      },
+    }),
+  }))
+  .add('with panel header but no refinements', () => ({
+    component: wrapWithHits({
+      template: `
+        <ais-panel header='Current refinements'>
+          <ais-current-refinements>
+          </ais-current-refinements>
+        </ais-panel>
+      `,
+    }),
+  }))
+  .add('with clearsQuery', () => ({
+    component: wrapWithHits({
+      template: `
+        <ais-panel header='Current refinements'>
+          <ais-current-refinements [clearsQuery]="true">
+          </ais-current-refinements>
+        </ais-panel>
+      `,
+      searchParameters: {
+        disjunctiveFacetsRefinements: { brand: ['Apple', 'Samsung'] },
+        disjunctiveFacets: ['brand'],
+        numericRefinements: { price: { '>=': [100] } },
+      },
+    }),
+  }));
