+++ examples/storybook/src/stories/Hits.stories.ts
@@ -1,0 +1,11 @@
+import { storiesOf } from '@storybook/angular';
+import { wrapWithHits } from '../wrap-with-hits';
+import meta from '../meta';
+
+storiesOf('Hits', module)
+  .addDecorator(meta)
+  .add('default', () => ({
+    component: wrapWithHits({
+      template: '<ais-hits></ais-hits>',
+    }),
+  }));
