+++ examples/storybook/src/stories/Pagination.stories.ts
@@ -1,0 +1,11 @@
+import { storiesOf } from '@storybook/angular';
+import { wrapWithHits } from '../wrap-with-hits';
+import meta from '../meta';
+
+storiesOf('Pagination', module)
+  .addDecorator(meta)
+  .add('default', () => ({
+    component: wrapWithHits({
+      template: '<ais-pagination></ais-pagination>',
+    }),
+  }));
