+++ examples/storybook/src/stories/RangeSlider.stories.ts
@@ -1,0 +1,14 @@
+import { storiesOf } from '@storybook/angular';
+import { wrapWithHits } from '../wrap-with-hits';
+import meta from '../meta';
+
+storiesOf('RangeSlider', module)
+  .addDecorator(meta)
+  .add('default', () => ({
+    component: wrapWithHits({
+      template: `
+      <ais-range-slider attribute="price">
+      </ais-range-slider>
+    `,
+    }),
+  }));
